LayoutTests/platform/mac-wk1/http/tests/security/mixedContent/redirect-https-to-http-image-secure-cookies-expected.txt

main frame - didFinishDocumentLoadForFrame
main frame - didStartProvisionalLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
main frame - didCommitLoadForFrame
CONSOLE MESSAGE: secureCookie=yes
main frame - didFinishDocumentLoadForFrame
CONSOLE MESSAGE: Cannot load image https://127.0.0.1:8443/resources/redirect.py?url=http://127.0.0.1:8080/security/resources/compass.jpg due to access control checks.
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This test opens a window that reads a secure cookie and then loads an insecure image (via a tricky redirect). The insecure image should be blocked because upgrading the request was not successful.
